+++ src/main/java/org/kpmp/spatialViewerDataset/SpatialViewerExternalLinkRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -13,5 +14,5 @@ public interface SpatialViewerExternalLinkRepository
     List<SpatialViewerExternalLinkDataset> findAll();
     
     @Query (value = "SELECT MAX(release_version) AS max FROM sv_link_v", nativeQuery = true)
-    public Double max();
+    public BigDecimal max();
 }
